requirements.txt
altair==4.2.0
async-generator==1.10
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.2
fonttools==4.37.3
gitdb==4.0.9
GitPython==3.1.27
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
Jinja2==3.1.2
jsonschema==4.16.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib
multitasking==0.0.11
numpy
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas
patsy==0.5.2
Pillow==9.2.0
plotly==5.10.0
plotly-express==0.4.1
protobuf==3.20.1
pyarrow==9.0.0
pycparser==2.21
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-futures==1.0.0
rich==12.5.1
scipy
selenium==4.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
statsmodels==0.13.2
streamlit
streamlit-option-menu==0.3.2
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
webdriver-manager==3.8.4
wsproto==1.2.0
yahooquery==2.2.15
yfinance==0.1.74
zipp==3.8.1
pymongo
SQLAlchemy
mysql
mysql-connector-python

